+++ src/internet/model/tcp-cubic.cc
@@ -189,6 +189,13 @@ TcpCubic::IncreaseWindow(Ptr<TcpSocketState> tcb, uint32_t segmentsAcked)
 {
     NS_LOG_FUNCTION(this << tcb << segmentsAcked);
 
+    if (!tcb->m_isCwndLimited)
+    {
+        NS_LOG_DEBUG("No increase because current cwnd " << tcb->m_cWnd
+                                                         << " is not limiting the flow");
+        return;
+    }
+
     if (tcb->m_cWnd < tcb->m_ssThresh)
     {
         if (m_hystart && tcb->m_lastAckedSeq > m_endSeq)