+++ src/internet/model/tcp-socket-state.h
@@ -207,6 +207,7 @@ class TcpSocketState : public Object
     Time m_minRtt{Time::Max()}; //!< Minimum RTT observed throughout the connection
 
     TracedValue<uint32_t> m_bytesInFlight{0};  //!< Bytes in flight
+    bool m_isCwndLimited{false};               //!< Whether throughput is limited by cwnd
     TracedValue<Time> m_lastRtt{Seconds(0.0)}; //!< Last RTT sample collected
 
     Ptr<TcpRxBuffer> m_rxBuffer; //!< Rx buffer (reordering buffer)